Introduce EventFlowBuilder
